stream_settings: Use `default_code_block_language`.
For syntax highlighting in code blocks, the
`default_code_block_language` of the stream is used in stream
messages. If this property is not set, we revert to the realm
default code block language (as earlier).
For direct messages, only the realm default is used.

- Added the field to the Stream type definition.
- Modified the messages/render handler to accept a default
code block language for rendering a message.

Fixes part of #17454